requirements.txt
pandas
numpy
scikit-learn
joblib
streamlit
matplotlib
openpyxl
Pillow